Configure RFM69 module via SPI on M0

+++ 02_SW/001_poc/000_M0plus_SPI_PoC/spi_master.h
@@ -32,4 +32,4 @@
 #define SPI_BAUD					50000			
 
 void SPI_init(void);
-uint8_t spiSend(uint8_t data);
+uint32_t spiSend(uint8_t data);
